+++ photutils/background/interpolators.py
@@ -1,7 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """
-This module defines classes to estimate the 2D background and background
-RMS in an image.
+This module defines interpolator classes for Background2D.
 """
 
 import numpy as np